separated multi_index printers into separate file
